requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
certifi
openai==1.82.0
langdetect==1.0.9
fpdf2==2.8.2
gunicorn==23.0.0
pytest==8.3.4
flake8==7.1.1
